extract router logic
